src/requirements.txt
python-telegram-bot==20.8
pydantic==2.6.3
python-dotenv==1.0.1
httpx~=0.26.0
pytest==8.0.2
pytest-asyncio==0.23.5 